requirements.txt
langgraph
langchain
langchain-community
openai>=1.0.0
python-dotenv